Stable version - Tests have run thousands of times, with only the "cannot-signal-condition exception"   bug occurring overy few thousand iterations. - Code still contains extensive manifest sanity checks. - Variable elision misses some opportunities. - Build sometimes leads to broken artifact that's unable to find All_Primitives.txt   during initialization.  Haven't investigated circumstances yet.

+++ avail/src/main/kotlin/avail/descriptor/tuples/TwoByteStringDescriptor.kt
@@ -128,7 +128,7 @@ class TwoByteStringDescriptor private constructor(
 				val offset = (subscript - 1) shl 2
 				for (i in offset + 1 .. min(self.tupleSize, offset + 4))
 				{
-					val c = self.shortSlot(this, i).toInt()
+					val c = self.shortSlot(this, i)
 					builder.appendCodePoint(
 						when
 						{
